synthetic/requirements.txt
blenderproc==2.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pip==25.0.1
progressbar==2.5
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pyyaml==6.0.2
requests==2.32.3
scipy==1.15.2
six==1.17.0
tqdm==4.67.1
urllib3==2.3.0
